# search-api/openapi.yaml
openapi: 3.1.0
info:
  title: search-api
  version: "1"
  description: |
    Search service backing the legacy `article-features` API via the ACL
    layer. Two routes share `POST /{collection}/_search*`:

      * `/{collection}/_search`     — F2+ contract: rich body carrying
        filters, currency, summaries, and pagination. Behaviour is being
        filled in across F3..F5; today it validates and returns an empty
        envelope.
      * `/{collection}/_search_v0`  — legacy ES-shaped contract preserved
        for the playground and dev-loop tooling. Deprecated; will be
        removed once the new contract is behaviour-complete.

    `collection` accepts a Milvus alias (e.g. `offers`); resolution to the
    versioned collection happens server-side.
servers:
  - url: http://localhost:8000
    description: Local dev / docker-compose default
security:
  - apiKeyHeader: []
  - apiKeyAuth: []
paths:
  /{collection}/_search:
    post:
      summary: Search (F2 contract — currently a stub).
      operationId: search
      description: |
        New contract per `issues/article-search-replacement-spec.md` §3
        and `issues/article-search-replacement-ftsearch-02-http-contract.md`.

        **F2 status**: validates the request shape and returns an empty
        envelope (`articles: []`, `metadata.hitCount: 0`, all summaries
        empty/null). F3 wires in scalar filtering; F4 fills in mode/sort/
        paging + accurate `hitCount`; F5 populates `summaries`.

        Pagination and sort are query-string parameters, not body fields.
        `pageSize` is capped at 500 server-side (legacy parity).
      parameters:
        - $ref: "#/components/parameters/Collection"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Sort"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchRequest"
            examples:
              hitsOnly:
                summary: Free-text search, hits only
                value:
                  searchMode: HITS_ONLY
                  query: "Bohrmaschine"
                  selectedArticleSources:
                    closedCatalogVersionIds: []
                    catalogVersionIdsOrderedByPreference: ["aaaaaaaa-1111-1111-1111-aaaaaaaaaaaa"]
                    sourcePriceListIds: ["bbbbbbbb-2222-2222-2222-bbbbbbbbbbbb"]
                    customerUploadedCoreArticleListSourceIds: []
                  currency: "EUR"
                  maxDeliveryTime: 5
                  coreSortimentOnly: false
                  closedMarketplaceOnly: false
              both:
                summary: Hits + summaries (faceted search)
                value:
                  searchMode: BOTH
                  query: "Akkubohrer"
                  selectedArticleSources:
                    closedCatalogVersionIds: []
                    catalogVersionIdsOrderedByPreference: ["aaaaaaaa-1111-1111-1111-aaaaaaaaaaaa"]
                    sourcePriceListIds: ["bbbbbbbb-2222-2222-2222-bbbbbbbbbbbb"]
                    customerUploadedCoreArticleListSourceIds: []
                  currency: "EUR"
                  maxDeliveryTime: 0
                  coreSortimentOnly: false
                  closedMarketplaceOnly: false
                  currentCategoryPathElements: ["Werkzeug", "Akku"]
                  requiredFeatures:
                    - name: "Spannung"
                      values: ["18V", "36V"]
                  priceFilter:
                    min: 0
                    max: 50000
                    currencyCode: "EUR"
                  summaries: ["VENDORS", "MANUFACTURERS", "FEATURES", "PRICES", "CATEGORIES"]
              summariesOnly:
                summary: Aggregations over the full filtered set, no articles
                value:
                  searchMode: SUMMARIES_ONLY
                  selectedArticleSources:
                    closedCatalogVersionIds: []
                    catalogVersionIdsOrderedByPreference: ["aaaaaaaa-1111-1111-1111-aaaaaaaaaaaa"]
                    sourcePriceListIds: []
                    customerUploadedCoreArticleListSourceIds: []
                  currency: "EUR"
                  maxDeliveryTime: 0
                  coreSortimentOnly: false
                  closedMarketplaceOnly: false
                  summaries: ["VENDORS"]
      responses:
        "200":
          description: Search executed (may have zero hits / empty summaries).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
              examples:
                empty:
                  summary: F2 stub response
                  value:
                    articles: []
                    summaries:
                      vendorSummaries: []
                      manufacturerSummaries: []
                      featureSummaries: []
                      pricesSummary: []
                      categoriesSummary: null
                      eClass5Categories: null
                      eClass7Categories: null
                      s2ClassCategories: null
                      eClassesAggregations: []
                    metadata:
                      page: 1
                      pageSize: 10
                      pageCount: 0
                      term: "Bohrmaschine"
                      hitCount: 0
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          description: Request body or query parameter failed Pydantic/FastAPI validation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "503":
          $ref: "#/components/responses/Throttled"
  /{collection}/_search_v0:
    post:
      deprecated: true
      summary: Legacy ES-shaped search (deprecated — use `/{collection}/_search`).
      description: |
        `collection` names the dense collection (e.g. `offers`). The
        BM25 codes companion collection (e.g. `offers_codes`) is
        server-configured via `MILVUS_CODES_COLLECTION`.

        ## Modes

        | mode                  | dense | bm25 | classifier | notes |
        | --------------------- | :---: | :--: | :--------: | ----- |
        | `vector`              | ✓     |      |            | dense ANN only |
        | `bm25`                |       | ✓    |            | BM25 over codes only |
        | `hybrid`              | ✓     | ✓    |            | RRF-fused, classifier off |
        | `hybrid_classified`   | ✓     | ✓    | ✓          | strict path for opaque IDs, else hybrid (default) |

        For `hybrid_classified`, the classifier may pick a strict
        BM25-only path with `strict_codes_limit` candidates. If that
        returns zero hits and `enable_fallback=1` (the default), the
        request falls through to a hybrid (dense + BM25, RRF-fused)
        execution.

        ## HNSW efSearch

        `num_candidates` maps to Milvus HNSW `ef`. It must be ≥ the dense
        leg's limit:
          * `vector` mode: `num_candidates ≥ k`
          * `hybrid` / `hybrid_classified`: `num_candidates ≥ dense_limit`

        ## Concurrency

        The server applies a hard in-process cap (`MAX_CONCURRENT_SEARCHES`,
        default 64). Excess requests get `503` with `Retry-After: 1`.
      operationId: searchV0
      parameters:
        - $ref: "#/components/parameters/Collection"
        - $ref: "#/components/parameters/Mode"
        - $ref: "#/components/parameters/K"
        - $ref: "#/components/parameters/DenseLimit"
        - $ref: "#/components/parameters/CodesLimit"
        - $ref: "#/components/parameters/StrictCodesLimit"
        - $ref: "#/components/parameters/RrfK"
        - $ref: "#/components/parameters/NumCandidates"
        - $ref: "#/components/parameters/EnableFallback"
        - $ref: "#/components/parameters/Debug"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchRequestV0"
            examples:
              simple:
                summary: A vanilla product-search query
                value:
                  query: "cordless drill 18v"
                  index: "offers-prod"
              opaqueId:
                summary: Opaque MPN — classifier will route to strict BM25
                value:
                  query: "GTB6-P5211"
                  index: "offers-prod"
      responses:
        "200":
          description: Search executed (may have zero hits).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponseV0"
              examples:
                hits:
                  summary: Two hits from the RRF-fused hybrid path
                  value:
                    hits:
                      - _index: "offers-prod"
                        _id: "offer:42"
                        _score: 0.0327
                        _source_leg: "rrf"
                      - _index: "offers-prod"
                        _id: "offer:99"
                        _score: 0.0312
                        _source_leg: "rrf"
                empty:
                  summary: Empty query short-circuits to no hits
                  value:
                    hits: []
                debug:
                  summary: With `?debug=1`
                  value:
                    hits:
                      - _index: "offers-prod"
                        _id: "offer:42"
                        _score: 0.0327
                        _source_leg: "rrf"
                    _debug:
                      path: "hybrid"
                      classifier_strict: false
                      fallback_fired: false
                      embed_ms: 12.4
                      dense_ms: 18.7
                      codes_ms: 9.1
                      dense_hits: 200
                      codes_hits: 18
                      params:
                        mode: "hybrid_classified"
                        k: 100
                        dense_limit: 200
                        codes_limit: 20
                        strict_codes_limit: 500
                        rrf_k: 60
                        num_candidates: null
                        enable_fallback: true
        "400":
          description: |
            Invalid query parameters. Examples: unknown `mode`, or
            `num_candidates` smaller than the dense-leg limit.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                badMode:
                  value:
                    detail: "invalid mode 'fuzzy'; expected one of: vector, bm25, hybrid, hybrid_classified"
                badNumCandidates:
                  value:
                    detail: "num_candidates (50) must be >= the dense leg's limit (200)"
        "401":
          description: Missing or invalid API key (only when `API_KEY` is configured server-side).
          headers:
            WWW-Authenticate:
              schema:
                type: string
              example: 'ApiKey realm="search-api"'
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                detail: "invalid or missing api key"
        "404":
          description: |
            The dense collection (or, for non-`vector` modes, the codes
            collection) does not exist in Milvus.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                noDense:
                  value:
                    detail: "Milvus dense collection 'offers' does not exist"
                noCodes:
                  value:
                    detail: "Milvus codes collection 'offers_codes' does not exist"
        "422":
          description: Request body or path parameter failed Pydantic/FastAPI validation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "503":
          description: |
            Server is at its in-process concurrency limit. Clients should
            back off and retry.
          headers:
            Retry-After:
              schema:
                type: string
              example: "1"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                detail: "search-api at concurrency limit"
...
components:
  securitySchemes:
    apiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
      description: Shared secret. Only enforced when `API_KEY` is set on the server.
    apiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Shared secret as `Authorization: ApiKey <key>`. Only enforced when
        `API_KEY` is set on the server. Either this header or `X-API-Key`
        is accepted.
  parameters:
    Collection:
      name: collection
      in: path
      required: true
      description: Milvus dense collection name (e.g. `offers`).
      schema:
        type: string
        minLength: 1
        maxLength: 255
        example: offers
    Mode:
      name: mode
      in: query
      description: Search execution mode. Defaults to `hybrid_classified`.
      schema:
        $ref: "#/components/schemas/Mode"
    K:
      name: k
      in: query
      description: Final top-N hits returned. Defaults to `SEARCH_TOP_K` (100).
      schema:
        type: integer
        minimum: 1
        example: 50
    DenseLimit:
      name: dense_limit
      in: query
      description: |
        Candidate pool size for the dense leg in `hybrid` /
        `hybrid_classified`. Ignored in `vector` (which uses `k` directly)
        and `bm25`.
      schema:
        type: integer
        minimum: 1
        default: 200
    CodesLimit:
      name: codes_limit
      in: query
      description: |
        Candidate pool size for the BM25 leg in `hybrid` /
        `hybrid_classified` (non-strict path) and in pure `bm25` mode.
      schema:
        type: integer
        minimum: 1
        default: 20
    StrictCodesLimit:
      name: strict_codes_limit
      in: query
      description: |
        Candidate pool size for the BM25 leg when `hybrid_classified`
        routes to the strict path (opaque ID / MPN queries).
      schema:
        type: integer
        minimum: 1
        default: 500
    RrfK:
      name: rrf_k
      in: query
      description: RRF `k` constant used to fuse dense and BM25 result lists.
      schema:
        type: integer
        minimum: 1
        default: 60
    NumCandidates:
      name: num_candidates
      in: query
      description: |
        HNSW `efSearch`. Must be ≥ the dense leg's limit (`k` in `vector`
        mode, otherwise `dense_limit`). Ignored by non-HNSW indexes.
      schema:
        type: integer
        minimum: 1
    EnableFallback:
      name: enable_fallback
      in: query
      description: |
        For `hybrid_classified` only. When `1` (default), a strict-path
        empty result falls through to hybrid execution. When `0`, the
        empty strict result is returned as-is.
      schema:
        $ref: "#/components/schemas/BoolFlag"
    Debug:
      name: debug
      in: query
      description: When `1`, include a `_debug` block in the response.
      schema:
        $ref: "#/components/schemas/BoolFlag"
    Page:
      name: page
      in: query
      description: 1-indexed page number.
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      description: Page size; capped server-side at 500 (legacy parity).
      schema:
        type: integer
        minimum: 0
        maximum: 500
        default: 10
    Sort:
      name: sort
      in: query
      description: |
        Repeatable sort clause of the form `<field>,<asc|desc>`.
        Supported fields: `name`, `price`, `articleId`.
      schema:
        type: array
        items:
          type: string
          example: "name,asc"
  responses:
    Unauthorized:
      description: Missing or invalid API key (only when `API_KEY` is configured server-side).
      headers:
        WWW-Authenticate:
          schema:
            type: string
          example: 'ApiKey realm="search-api"'
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            detail: "invalid or missing api key"
    Throttled:
      description: |
        Server is at its in-process concurrency limit. Clients should
        back off and retry.
      headers:
        Retry-After:
          schema:
            type: string
          example: "1"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            detail: "search-api at concurrency limit"
  schemas:
...
    SearchRequestV0:
      type: object
      required: [query, index]
      properties:
        query:
          type: string
          description: |
            User query. Whitespace is trimmed; an empty query
            short-circuits to `{ "hits": [] }` without contacting Milvus
            or the embedder.
          example: cordless drill 18v
        index:
          type: string
          description: |
            Opaque label echoed back as `_index` on each hit. Lets
            callers tag results with their logical index name without
            the server having to know about it.
          example: offers-prod
        category:
          type: string
          nullable: true
          description: |
            **Reserved.** Currently accepted but unused by the server.
            Kept on the contract for forward compatibility.
          default: null
    SearchResponseV0:
      type: object
      required: [hits]
      properties:
        hits:
          type: array
          items:
            $ref: "#/components/schemas/HitV0"
        _debug:
          $ref: "#/components/schemas/DebugV0"
          description: Present only when `?debug=1`.
...
    SearchMode:
      type: string
      enum: [HITS_ONLY, SUMMARIES_ONLY, BOTH]
...
    FeatureFilter:
      type: object
      additionalProperties: false
      required: [name, values]
      properties:
        name: { type: string }
        values:
          type: array
          items: { type: string }
...
    SearchRequest:
      type: object
      additionalProperties: false
      required: [searchMode, selectedArticleSources, currency]
      description: |
        ftsearch search request. Two intentional deltas vs. legacy §3:
          * `searchArticlesBy` removed — only `STANDARD` is supported and
            the ACL rejects other values upstream (§2.1).
          * `explain` removed — the ACL stubs `"N/A"` per §2.2.
      properties:
        query:
          type: string
          nullable: true
        searchMode:
          $ref: "#/components/schemas/SearchMode"
        selectedArticleSources:
          $ref: "#/components/schemas/SelectedArticleSources"
        articleIdsFilter:
          type: array
          items: { type: string }
        vendorIdsFilter:
          type: array
          items: { type: string }
        manufacturersFilter:
          type: array
          items: { type: string }
        maxDeliveryTime:
          type: integer
          minimum: 0
          default: 0
        requiredFeatures:
          type: array
          items:
            $ref: "#/components/schemas/FeatureFilter"
        priceFilter:
          $ref: "#/components/schemas/PriceFilter"
          nullable: true
        accessoriesForArticleNumber:
          type: string
          nullable: true
        sparePartsForArticleNumber:
          type: string
          nullable: true
        similarToArticleNumber:
          type: string
          nullable: true
        currentCategoryPathElements:
          type: array
          items: { type: string }
        currentEClass5Code:
          type: integer
          nullable: true
        currentEClass7Code:
          type: integer
          nullable: true
        currentS2ClassCode:
          type: integer
          nullable: true
        coreSortimentOnly:
          type: boolean
          default: false
        closedMarketplaceOnly:
          type: boolean
          default: false
        summaries:
          type: array
          items:
            $ref: "#/components/schemas/SummaryKind"
        coreArticlesVendorsFilter:
          type: array
          items: { type: string }
        blockedEClassVendorsFilters:
          type: array
          items:
            $ref: "#/components/schemas/BlockedEClassVendorsFilter"
        currency:
          type: string
          pattern: "^[A-Z]{3}$"
        eClassesFilter:
          type: array
          items: { type: integer }
        eClassesAggregations:
          type: array
          items:
            $ref: "#/components/schemas/EClassesAggregation"
        s2ClassForProductCategories:
          type: boolean
          default: false
...
    SearchResponse:
      type: object
      additionalProperties: false
      required: [articles, summaries, metadata]
      properties:
        articles:
          type: array
          items:
            $ref: "#/components/schemas/Article"
        summaries:
          $ref: "#/components/schemas/Summaries"
        metadata:
          $ref: "#/components/schemas/Metadata"
    Error:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
    ValidationError:
      type: object
      description: FastAPI / Pydantic validation error envelope.
      required: [detail]
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  oneOf:
                    - type: string
                    - type: integer
              msg:
                type: string
              type:
                type: string